Fix "strategic" alarm warning
I'd avoid this word as much as possible; otherwise we may find
ourselves one day giving PowerPoint presentations on the goods
of OOP to an entirely managerial audience.

Hyperboles aside, we use "critical" in abstract v2, which btw isn't
quite the same as "strategic", and it's nice to keep them consistent.

Signed-off-by: Emilio G. Cota <cota@braap.org>